Ai_to_db/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
firebase-admin==6.5.0
groq==1.0.0
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests>=2.31.0
certifi>=2024.0.0
charset-normalizer>=3.0.0
idna>=3.4
urllib3>=2.0.0
